RUM-12062: Remove reflection
